More work on server logic + small other things

+++ server/src/main/java/imag/quizz/server/game/Server.java
@@ -7,7 +7,7 @@ public final class Server extends Peer {
 
     private boolean isLeader;
 
-    public Server(final int id, final int port) {
+    public Server(final long id, final int port) {
         super(id, port);
     }
